Simulator final

+++ ProcessScheduler/Processors/FCFS.cpp
@@ -6,33 +6,40 @@ void FCFS::SigKill(Process* P){
 }
 
 void FCFS::ScheduleAlgo(){
-    Process* nR;
-    nR = RDY.getHead()->getItem();
-    setRUN(nR);
-    RDY.DeleteFirst();
+    Node<Process*>* nR;
+    nR = RDY.getHead();
+    if (nR) {
+        setRUN(nR->getItem());
+        RDY.DeleteFirst();
+    }
+    else {
+        setRUN(nullptr);
+    }
 }
 
 
   
 
-void FCFS::removeFromReady(int pid)
+Process* FCFS::removeFromReady(int pid)
 {
     Node<Process*>* head = RDY.getHead();
     Process* to_remove;
+    if (!head) return nullptr;
 
-    while (head)
+    while (head->getNext())
     {
         if (head->getItem()->find_by_pid(pid))
         {
             to_remove = head->getItem();
             RDY.DeleteNode(to_remove);
-            return; //assuming 2 processes cant have the same pid
+            return to_remove; //assuming 2 processes cant have the same pid
         }
         else 
         {
             head = head->getNext();
         }
     }
+    return nullptr;
     
 }
 
@@ -40,14 +47,14 @@ FCFS::FCFS(){
 
 }
 
-void FCFS::moveToRDY(Process* NewProcess)
+void FCFS::moveToRDY(Process* const& NewProcess)
 {
-//    RDY.enqueue(NewProcess);
+    RDY.InsertEnd(NewProcess);
 }
 
 ostream& operator<<(ostream& os, const FCFS& prcsr)
 {
-    os << "[FCFS]: " << "prcsr.RDY.getCount()" << "RDY: ";
+    os << "[FCFS]: " << prcsr.RDY.getCount() << " RDY: ";
     prcsr.RDY.PrintList();
     return os;
 }
